[SCB-2029] tiny improve to make extend filter features easier

+++ core/src/main/java/org/apache/servicecomb/core/exception/converter/IllegalArgumentExceptionConverter.java
@@ -24,7 +24,7 @@ import org.apache.servicecomb.core.exception.ExceptionConverter;
 import org.apache.servicecomb.swagger.invocation.exception.InvocationException;
 
 public class IllegalArgumentExceptionConverter implements ExceptionConverter<IllegalArgumentException> {
-  public static final short ORDER = Short.MAX_VALUE;
+  public static final int ORDER = Short.MAX_VALUE;
 
   @Override
   public int getOrder() {
